pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[project]
name = "omnidaemon"
dynamic = ["version"]
description = "Universal Event-Driven Runtime Engine for AI Agents"
readme = "README.md"
authors = [{ name = "Abiola Adeshina", email = "abiolaadedayo1993@gmail.com" }]
requires-python = ">=3.10"

keywords = [
    "ai",
    "agent",
    "automation",
    "event-driven",
    "runtime",
    "redis-streams",
    "kafka",
    "background-tasks",
    "microservices",
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Distributed Computing",
]

# Core runtime dependencies only
dependencies = [
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiohttp>=3.9.0",
    "python-decouple>=3.8",
    "psutil>=7.1.3",
    "pip>=25.3",
]

[project.scripts]
omnidaemon = "omnidaemon.cli.main:app"

[project.urls]
Homepage = "https://github.com/omnirexflora-labs/OmniDaemon"
Repository = "https://github.com/omnirexflora-labs/OmniDaemon"
Documentation = "https://docs-omnidaemon.omnirexfloralabs.com/docs"
Issues = "https://github.com/omnirexflora-labs/OmniDaemon/issues"
Changelog = "https://github.com/omnirexflora-labs/OmniDaemon/releases"

[tool.hatch.build.targets.wheel]
packages = ["src/omnidaemon"]

[tool.hatch.build]
packages = ["src/omnidaemon"]
include = [
    "LICENSE",
    "README.md"
]

[tool.mypy]
mypy_path = ["src"]
explicit_package_bases = true
ignore_missing_imports = true


[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests that use asyncio",
    "integration: marks integration tests",
]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/omnidaemon"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.uv.workspace]
members = [
    "omnidaemon_supervisor_test_dir/omnicore_agent", "omnidaemon_supervisor_test_dir", "examples/google_adk_with_supervisor",
]

[dependency-groups]
# Development dependencies
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.11.7",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "pre-commit>=4.2.0",
    "hatch>=1.14.1",
    "twine>=6.1.0",
]


# Example dependencies (optional)
examples = [
    "omnicoreagent>=0.2.10",
    "google-adk>=0.1.0",
    "litellm>=1.0.0",
]
